Fix: signed integer overflow in credit card number generation.
Reading a credit card fabricates a fake card number from the
object's o_id. One of the digit expressions multiplied o_id in
int before reducing mod 10, so once o_id passed INT_MAX / 7
(~307 million object allocations) the multiply overflowed signed
int - undefined behavior that aborts under UBSan. A long game
reached an o_id of ~347 million and crashed on the next credit
card read. The adjacent expression already promotes to long
before multiplying; this one did not.

* read.c: in doread(), compute the final card-number digit as
  (long) o_id * 7L before the mod, matching the sibling
  expression above it, then cast the single-digit result back to
  int for the format argument. Output is unchanged for all valid
  o_id.

+++ src/read.c
@@ -414,7 +414,7 @@ doread()
               ((((long) scroll->o_id * 499L) % 899999L) + 100000L),
               ((int) scroll->o_id % 10),
               (!((int) scroll->o_id % 3)),
-              (((int) scroll->o_id * 7) % 10),
+              (int) (((long) scroll->o_id * 7L) % 10L),
               (flags.verbose || Blind) ? "." : "");
         if (!u.uconduct.literate++)
             livelog_write_string(LL_CONDUCT,
